abstracao card theme dark
