+++ src/pages/comentarios.html
@@ -15,7 +15,7 @@
 	</head>
     
 <body>
-  		<nav class="navbar navbar-expand-lg navbar-light bg-light">
+  		<nav class="navbar navbar-expand-lg navbar-dark bg-primary">
 			<a class="navbar-brand" href="./index.html">Curso Lero Lero</a>
 			<button
 				class="navbar-toggler"
@@ -54,7 +54,7 @@
 				</ul>
 			</div>
 		</nav>
-    <div class="container">
+    <main class="container-fluid bg-dark pt-4">
     <div class="card-deck">
 
         <div class="card" style="width: 18rem;">
@@ -139,7 +139,7 @@
 		<script src="../../lib/popper/popper.min.js"></script>
 		<script src="../../lib/bootstrap/bootstrap.min.js"></script>
 		<!-- Script Projeto -->
-		<script type="module" src="../scripts/app.js"></script>
+		<script type="module" src="../scripts/comentarios.js"></script>
 </body>
 
 </html>
