+++ src/pages/index.html
@@ -13,7 +13,7 @@
 		<title>Curso Lero Lero - Início</title>
 	</head>
 	<body>
-		<nav class="navbar navbar-expand-lg navbar-light bg-light">
+		<nav class="navbar navbar-expand-lg navbar-dark bg-primary">
 			<a class="navbar-brand" href="./index.html">Curso Lero Lero</a>
 			<button
 				class="navbar-toggler"
@@ -52,25 +52,31 @@
 				</ul>
 			</div>
 		</nav>
-		<main class="container">
-			<div class="jumbotron d-flex flex-column">
-				<h1 class="display-4 d-none d-sm-block">Curso Lero Lero</h1>
-				<h1 class="d-sm-none">Curso Lero Lero</h1>
-				<p class="lead mt-2">
-					Preparando você para aproveitar a vida, contemplando o espaço e
-					apreciando o ócio.
-				</p>
-				<hr class="my-4" />
-				<p>
-					Não faça hoje o que você pode deixar para amanhã.
-				</p>
-				<a
-					class="btn btn-primary btn-lg align-self-md-end"
-					href="./sobre.html"
-					role="button"
-					>Saiba mais</a
-				>
-			</div>
+		<main class="container-fluid bg-dark pt-4 d-flex justify-content-center">
+			<section class="row">
+				<article class="col">
+					<div
+						class="jumbotron d-flex flex-column bg-dark text-white border border-secondary"
+					>
+						<h1 class="display-4 d-none d-sm-block">Curso Lero Lero</h1>
+						<h1 class="d-sm-none">Curso Lero Lero</h1>
+						<p class="lead mt-2">
+							Preparando você para aproveitar a vida: contemplando o espaço e
+							apreciando o ócio.
+						</p>
+						<hr class="my-4" />
+						<p>
+							Não faça hoje o que você pode deixar para amanhã.
+						</p>
+						<a
+							class="btn btn-primary btn-lg align-self-md-end"
+							href="./sobre.html"
+							role="button"
+							>Saiba mais</a
+						>
+					</div>
+				</article>
+			</section>
 		</main>
 		<!-- Bibliotecas -->
 		<script src="../../lib/jquery/jquery-3.4.1.min.js"></script>
