+++ src/pages/instrutores.html
@@ -13,7 +13,7 @@
 		<title>Curso Lero Lero - Instrutores</title>
 	</head>
 	<body>
-		<nav class="navbar navbar-expand-lg navbar-light bg-light">
+		<nav class="navbar navbar-expand-lg navbar-dark bg-primary">
 			<a class="navbar-brand" href="./index.html">Curso Lero Lero</a>
 			<button
 				class="navbar-toggler"
@@ -52,7 +52,7 @@
 				</ul>
 			</div>
 		</nav>
-		<main class="container">
+		<main class="container-fluid bg-dark">
 			<section class="row instrutores"></section>
 		</main>
 		<!-- Bibliotecas -->
