+++ src/pages/login.html
@@ -14,7 +14,7 @@
 	</head>
 
 	<body>
-		<nav class="navbar navbar-expand-lg navbar-light bg-light">
+		<nav class="navbar navbar-expand-lg navbar-dark bg-primary">
 			<a class="navbar-brand" href="./index.html">Curso Lero Lero</a>
 			<button
 				class="navbar-toggler"
@@ -53,31 +53,35 @@
 				</ul>
 			</div>
 		</nav>
-		<div class="container">
-			<div class="wrapper">
-				<form class="form-signin">
-					<h2 class="form-signin-heading">Please login</h2>
-					<input
-						type="text"
-						class="form-control"
-						name="username"
-						placeholder="Email Address"
-						required=""
-						autofocus=""
-					/>
-					<input
-						type="password"
-						class="form-control"
-						name="password"
-						placeholder="Password"
-						required=""
-					/>
-					<button class="btn btn-lg btn-primary btn-block" type="submit">
-						Login
-					</button>
-				</form>
-			</div>
-		</div>
+		<main class="container-fluid bg-dark pt-4 d-flex justify-content-center">
+			<section class="row">
+				<article class="col">
+					<form class="text-light border border-secondary rounded px-4 py-4">
+						<h2 class="pb-3">Login</h2>
+						<div class="form-group">
+							<label for="inputEmail">Email</label>
+							<input
+								type="email"
+								class="form-control"
+								id="inputEmail"
+								aria-describedby="emailHelp"
+								autofocus
+							/>
+							<small id="emailHelp" class="form-text text-muted"
+								>Vamos vender seus dados pra quem pagar mais.</small
+							>
+						</div>
+						<div class="form-group">
+							<label for="inputPassword">Senha</label>
+							<input type="password" class="form-control" id="inputPassword" />
+						</div>
+						<button type="submit" class="btn btn-primary">
+							Submit
+						</button>
+					</form>
+				</article>
+			</section>
+		</main>
 		<!-- Bibliotecas -->
 		<script src="../../lib/jquery/jquery-3.4.1.min.js"></script>
 		<script src="../../lib/jquery/jquery.validate.min.js"></script>
