+++ src/pages/sobre.html
@@ -13,7 +13,7 @@
 		<title>Curso Lero Lero - Sobre</title>
 	</head>
 	<body>
-		<nav class="navbar navbar-expand-lg navbar-light bg-light">
+		<nav class="navbar navbar-expand-lg navbar-dark bg-primary">
 			<a class="navbar-brand" href="./index.html">Curso Lero Lero</a>
 			<button
 				class="navbar-toggler"
