+++ src/pages/tabelas.html
@@ -13,7 +13,7 @@
 		<title>Curso Lero Lero - Tabelas</title>
 	</head>
 	<body>
-		<nav class="navbar navbar-expand-lg navbar-light bg-light">
+		<nav class="navbar navbar-expand-lg navbar-dark bg-primary">
 			<a class="navbar-brand" href="./index.html">Curso Lero Lero</a>
 			<button
 				class="navbar-toggler"
@@ -52,7 +52,9 @@
 				</ul>
 			</div>
 		</nav>
-		<main class="container">
+		<main
+			class="container-fluid text-light bg-dark pt-4 d-flex flex-column justify-content-center align-items-center"
+		>
 			<form method="POST" id="formulario" class="mx-3">
 				<h1>Formulário</h1>
 				<p>
