+++ src/scripts/CardInstrutor.js
@@ -1,6 +1,6 @@
-export default function Card({ title, desc, src }) {
-	const card = `<article class="col">
-  <div class="card" style="width: 18rem;">
+export default function CardInstrutor({ title, desc, src }) {
+	const card = `<article class="col d-flex justify-content-center">
+  <div class="card bg-dark text-light border border-tertiary my-3" style="width: 18rem;">
     <img src="${src}" class="card-img-top" alt="${title}" />
     <div class="card-body">
       <h5 class="card-title">${title}</h5>
