+++ src/styles/app.css
@@ -2,20 +2,6 @@
 	color: blue;
 }
 
-body,
-.container {
-	background: #eee;
-}
-
-.wrapper {
-	margin-top: 80px;
-	margin-bottom: 80px;
-}
-
-.form-signin {
-	max-width: 380px;
-	padding: 15px 35px 45px;
-	margin: 0 auto;
-	background-color: #fff;
-	border: 1px solid rgba(0, 0, 0, 0.1);
+.container-fluid {
+	min-height: 93vh;
 }
